Move Bills types to shared folder
