+++ components/bill/types.ts
@@ -1,4 +1,4 @@
-import { Bill } from "../db"
+import { Bill } from "common/bills/types"
 
 export type BillProps = { bill: Bill }
 export type { BillTracker } from "functions/src/analysis/types"
