+++ components/links.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link"
 import { forwardRef, PropsWithChildren } from "react"
-import { BillTopic, CurrentCommittee } from "../functions/src/bills/types"
-import { Bill, MemberContent } from "./db"
+import { Bill, BillTopic, CurrentCommittee } from "common/bills/types"
+import { MemberContent } from "./db"
 import { formatBillId } from "./formatting"
 import { TFunction } from "next-i18next"
 import { Testimony } from "common/testimony/types"
