+++ functions/src/testimony/publishTestimony.ts
@@ -2,7 +2,7 @@ import { DocumentReference, DocumentSnapshot } from "@google-cloud/firestore"
 import { https, logger } from "firebase-functions"
 import { nanoid } from "nanoid"
 import { Record } from "runtypes"
-import { Bill } from "../bills/types"
+import { Bill } from "../../../common/bills/types"
 import { checkAuth, checkRequest, DocUpdate, fail, Id } from "../common"
 import { db, FieldValue } from "../firebase"
 import { Timestamp } from "../../../common/types"
