+++ pages/bills/[court]/[billId].tsx
@@ -2,7 +2,7 @@ import { dbService } from "components/db/api"
 import { GetServerSideProps } from "next"
 import { z } from "zod"
 import { BillDetails } from "../../../components/bill"
-import { Bill } from "../../../components/db"
+import { Bill } from "common/bills/types"
 import { createPage } from "../../../components/page"
 import { usePublishService } from "../../../components/publish/hooks"
 import { serverSideTranslations } from "next-i18next/serverSideTranslations"
