+++ scripts/firebase-admin/generateBillHistory.ts
@@ -1,7 +1,7 @@
 import { Timestamp } from "../../common/types"
 import { Record, String, Number } from "runtypes"
 import { Script } from "./types"
-import { BillHistoryAction } from "../../functions/src/bills/types"
+import { BillHistoryAction } from "../../common/bills/types"
 import { FieldValue } from "functions/src/firebase"
 
 const Args = Record({
