+++ stories/organisms/SubmitTestimonyForm.stories.tsx
@@ -84,7 +84,8 @@ export const FormStory = {
           fetchedAt: Timestamp.fromDate(new Date()),
           history: [],
           currentCommittee: undefined,
-          city: undefined
+          city: undefined,
+          similar: []
         }}
         synced={false}
       />
